src/backend/requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-Session==0.5.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.20.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
nltk==3.9.2
numpy
openai==2.8.1
packaging==25.0
pandas
pyBKT==1.4.1
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.8.2
pytz==2025.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.29.0
scikit-learn
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
textblob==0.17.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.0.1
python-dotenv
